# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Wallboard Backend API
  description: |
    Backend API für die Wallboard/Dashboard-Anwendung mit Unterstützung für:
    - Dashboard- und Widget-Verwaltung
    - Google Calendar Integration  
    - Spotify Integration
    - SmartThings IoT Integration
    - Real-time Updates über Socket.IO
  version: 1.0.0
  contact:
    name: API Support
    url: https://screen-board.com
  license:
    name: Proprietary
servers:
  - url: https://api.screen-board.com/api
    description: Produktionsserver
  - url: http://localhost:5000/api
    description: Entwicklungsserver
security:
  - bearerAuth: []
...
paths:
...
  /dashboard/{dashboardId}/widgets:
    get:
      tags:
        - Widget
      summary: Alle Widgets eines Dashboards
      description: Ruft alle Widgets für ein spezifisches Dashboard ab
      operationId: getWidgetsByDashboardId
      parameters:
        - name: dashboardId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Liste der Widgets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Widget'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /dashboard/{dashboardId}/widgets/{id}:
    get:
      tags:
        - Widget
      summary: Widget abrufen
      description: Ruft ein spezifisches Widget ab
      operationId: getWidgetById
      parameters:
        - name: dashboardId
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Widget erfolgreich abgerufen
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Widget'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: Widget nicht gefunden
    
    put:
      tags:
        - Widget
      summary: Widget aktualisieren
      description: Aktualisiert ein bestehendes Widget
      operationId: updateWidget
      parameters:
        - name: dashboardId
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WidgetUpdate'
      responses:
        '200':
          description: Widget erfolgreich aktualisiert
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Widget'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: Widget nicht gefunden

  # Layout Endpoints
...
  /users:
    get:
      tags:
        - User
      summary: Alle Benutzer abrufen
      description: Ruft alle registrierten Benutzer ab
      operationId: getAllUsers
      responses:
        '200':
          description: Liste der Benutzer
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    
    post:
      tags:
        - User
      summary: Benutzer erstellen
      description: Erstellt einen neuen Benutzer
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '201':
          description: Benutzer erfolgreich erstellt
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Ungültige Eingabedaten
  /users/{id}:
    get:
      tags:
        - User
      summary: Benutzer nach ID abrufen
      description: Ruft einen spezifischen Benutzer ab
      operationId: getUserById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Benutzer erfolgreich abgerufen
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: Benutzer nicht gefunden

  # Google Calendar Endpoints
...
  /spotify/setActiveDevice:
    post:
      tags:
        - Spotify
      summary: Aktives Gerät setzen
      description: Setzt das aktive Spotify-Wiedergabegerät
      operationId: setActiveDevice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - dashboardId
                - deviceId
              properties:
                dashboardId:
                  type: string
                deviceId:
                  type: string
      responses:
        '200':
          description: Gerät erfolgreich gesetzt
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  # SmartThings Endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Auth0 JWT Token. Alle API-Endpunkte (außer OAuth-Callbacks und Webhooks) 
        erfordern einen gültigen JWT-Token im Authorization-Header.
  schemas:
...
    Widget:
      type: object
      properties:
        _id:
          type: string
        dashboardId:
          type: string
        type:
          type: string
          enum:
            - datetime
            - weather
            - music
            - events
            - notes
            - qrcodes
            - remarks
            - iot
            - news
        config:
          type: object
          description: Widget-spezifische Konfiguration
        position:
          type: object
          properties:
            x:
              type: integer
            y:
              type: integer
            w:
              type: integer
            h:
              type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - _id
        - dashboardId
        - type
    WidgetUpdate:
      type: object
      properties:
        type:
          type: string
        config:
          type: object
        position:
          type: object
...
    User:
      type: object
      properties:
        _id:
          type: string
        auth0Id:
          type: string
          description: Auth0-Benutzer-ID
        email:
          type: string
          format: email
        name:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - _id
        - auth0Id
        - email
    UserCreate:
      type: object
      properties:
        auth0Id:
          type: string
        email:
          type: string
          format: email
        name:
          type: string
      required:
        - auth0Id
        - email
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
      required:
        - error
        - message
  responses:
    UnauthorizedError:
      description: Authentifizierung fehlgeschlagen oder Token ungültig
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Invalid token"
            message: "Unauthorized"